event_bot/requirements.txt
python-telegram-bot==20.6
google-generativeai==0.4.1
requests
python-dotenv
